openai_agent_retrieval/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.3
python-dotenv==1.0.1
qdrant-client==1.16.1
openai==1.55.0
cohere==4.27
requests==2.32.5
pytest==8.3.5
httpx==0.28.1